[EDIFIKANA] #344 Renaming FileManager to FileService based on PR feedback
