+++ edifikana/front-end/shared-app/src/androidMain/kotlin/com/cramsan/edifikana/client/lib/di/ManagerPlatformModule.android.kt
@@ -1,7 +1,7 @@
 package com.cramsan.edifikana.client.lib.di
 
-import com.cramsan.edifikana.client.lib.managers.FileManager
-import com.cramsan.edifikana.client.lib.managers.FileManagerImpl
+import com.cramsan.edifikana.client.lib.service.FileService
+import com.cramsan.edifikana.client.lib.managers.FileServiceImpl
 import com.cramsan.edifikana.client.lib.service.AndroidDownloadStrategy
 import com.cramsan.edifikana.client.lib.service.DownloadStrategy
 import com.cramsan.edifikana.client.lib.utils.IODependencies
@@ -16,7 +16,7 @@ internal actual val ManagerPlatformModule = module {
         AndroidDownloadStrategy(get())
     }
 
-    single<FileManager> {
-        FileManagerImpl(get())
+    single<FileService> {
+        FileServiceImpl(get())
     }
 }
